requirements.txt
flask
joblib
pandas
kafka-python
